--- test/earmark_helpers_tests/inline_code/opens_inline_code_test.exs
@@ -1,22 +1,0 @@
-
-defmodule Helpers.InlineCodeTest.OpensInlineCodeTest do
-  use ExUnit.Case
-  import Earmark.Helpers.InlineCodeHelpers, only: [opens_inline_code: 1]
-  [
-    { "empty line -> not a pending inline code", "", {nil, 0} },
-    { "no backquotes -> not a pending inline code", "Hello World", {nil, 0} },
-    { "closed backquotes --> not a pending inline code", "`a ``` b`c", {nil, 0} },
-    { "pair of backquotes, nested structs -> not pending", "`1 `` 2` `` ` `` >`< `` >`<", {nil, 0}},
-    { "pair of backquotes, some escapes -> not pending", "`1 `` 2` `` ` `` >`< `` \\`>`<", {nil, 0}},
-
-    { "one single backquote -> pending(`)", "`", {"`", 42} },
-    { "one double backquote -> pending(``)", "``", {"``", 42} },
-    { "triple backquote after some text -> pending(```)", "alpha```", {"```", 42} },
-    { "single backquote in between -> pending(`)", "`1 `` 2` `` ` `` >`< ``", {"`", 42}},
-    { "single backquote in between, some escapes -> pending(`)", "`1 `` 2` `` ` `` \\`>`< ``", {"`", 42}}
-  ] |> Enum.each(fn { description, line, result } -> 
-           test(description) do
-             assert opens_inline_code(%{line: unquote(line), lnb: 42}) == unquote(result)
-           end
-         end)
-end
